requirements.txt
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
asgiref==3.5.1
Django==3.2
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-pwa==1.0.10
django-widget-tweaks==1.4.12
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
django-environ==0.10.0  # https://github.com/joke2k/django-environ
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
typing-extensions==4.7.1

django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.3.0  # https://github.com/jazzband/django-redis
psycopg2==2.9.3


python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
django-debug-toolbar==3.2.1
cloudinary==1.34.0
django-cloudinary-storage==0.3.0
django-ckeditor==6.5.1